Update installer.py not to destroy the original PYTHONPATH for installer subprocesses
When invoking the extension installer subprocess, the code used to simply overwrite the PYTHONPATH env var, which could potentially mess with the python runtime in some environments, basically breaking the extension installation step. 

Changing this behavior to simply "add" the current working directory to the path and not destroy its other entries already there, would solve the problem and allows the whole application run just fine.