Remove conditional Hires sampler omission
Fixes restoring parameters when your default hires sampler isn't "Same as primary". In those cases, only the main sampler would be updated because there's no value stored for the hires sampler.

While technically you *could* assume that "Same as primary" is meant when other hires settings are loaded without a value for the sampler, it would be a lot more complicated to actually implement. Honestly, it's just better and probably more "expected behavior" to always include the information when using Hires.